hyflux-ppt-automation/requirements.txt
python-pptx==0.6.23
PyYAML==6.0.1
Pillow==10.1.0
